Release v0.2.6: DuckDB Engine Update & Code Quality Improvements
🚀 Major Updates:
- Updated DuckDB engine to v2.3.3+ for enhanced performance
- Updated Apache Arrow to v18.4.0 for improved data interchange
- Comprehensive dependency updates to latest stable versions

🔧 Code Quality & Organization:
- Reorganized test files into dedicated test/ directory
- Fixed all golangci-lint issues (0 linting errors)
- Added constants for repeated string literals (goconst)
- Converted if-else chains to switch statements (gocritic)
- Updated deprecated driver methods to context-aware versions (staticcheck)
- Removed unused functions and improved code maintainability

📦 Package Structure:
- Updated test files to use proper package duckdb_test structure
- Resolved function name conflicts across test files
- Enhanced type references with proper package prefixes
- Improved module organization with clean import paths

✅ Validation:
- All 100+ tests pass with updated structure
- Zero golangci-lint issues across entire codebase
- Full backward compatibility maintained
- Complete CRUD, array, and extension functionality verified

Breaking Changes: None
Compatibility: Go 1.24+, DuckDB v2.3.3+, GORM v1.25.12
